Refactor Kaboom classes
