+++ src/kaboom.ts
@@ -151,6 +151,13 @@ import {
 	TweenController,
 } from "./types"
 
+import { Texture } from "./classes/Texture"
+import { SpriteData } from "./classes/SpriteData"
+import { SoundData } from "./classes/SoundData"
+import { Asset } from "./classes/Asset"
+import { AssetBucket } from "./classes/AssetBucket"
+import { Collision } from "./classes/Collision"
+
 import FPSCounter from "./fps"
 import Timer from "./timer"
 
@@ -529,88 +536,6 @@ export default (gopt: KaboomOpt = {}): KaboomCtx => {
 			preserveDrawingBuffer: true,
 		})
 
-	class Texture {
-
-		glTex: WebGLTexture
-		width: number
-		height: number
-
-		constructor(w: number, h: number, opt: TextureOpt = {}) {
-
-			this.glTex = gl.createTexture()
-			gc.push(() => this.free())
-			this.bind()
-
-			if (w && h) {
-				gl.texImage2D(
-					gl.TEXTURE_2D,
-					0, gl.RGBA,
-					w,
-					h,
-					0,
-					gl.RGBA,
-					gl.UNSIGNED_BYTE,
-					null,
-				)
-			}
-
-			this.width = w
-			this.height = h
-
-			const filter = (() => {
-				switch (opt.filter ?? gopt.texFilter) {
-					case "linear": return gl.LINEAR
-					case "nearest": return gl.NEAREST
-					default: return gl.NEAREST
-				}
-			})()
-
-			const wrap = (() => {
-				switch (opt.wrap) {
-					case "repeat": return gl.REPEAT
-					case "clampToEdge": return gl.CLAMP_TO_EDGE
-					default: return gl.CLAMP_TO_EDGE
-				}
-			})()
-
-			gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MIN_FILTER, filter)
-			gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MAG_FILTER, filter)
-			gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_S, wrap)
-			gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_T, wrap)
-			this.unbind()
-
-		}
-
-		static fromImage(img: TexImageSource, opt: TextureOpt = {}): Texture {
-			const tex = new Texture(0, 0, opt)
-			tex.bind()
-			gl.texImage2D(gl.TEXTURE_2D, 0, gl.RGBA, gl.RGBA, gl.UNSIGNED_BYTE, img)
-			tex.width = img.width
-			tex.height = img.height
-			tex.unbind()
-			return tex
-		}
-
-		update(x: number, y: number, img: TexImageSource) {
-			this.bind()
-			gl.texSubImage2D(gl.TEXTURE_2D, 0, x, y, gl.RGBA, gl.UNSIGNED_BYTE, img)
-			this.unbind()
-		}
-
-		bind() {
-			gl.bindTexture(gl.TEXTURE_2D, this.glTex)
-		}
-
-		unbind() {
-			gl.bindTexture(gl.TEXTURE_2D, null)
-		}
-
-		free() {
-			gl.deleteTexture(this.glTex)
-		}
-
-	}
-
 	const gfx = (() => {
 
 		const defShader = makeShader(DEF_VERT, DEF_FRAG)
@@ -618,7 +543,8 @@ export default (gopt: KaboomOpt = {}): KaboomCtx => {
 		// a 1x1 white texture to draw raw shapes like rectangles and polygons
 		// we use a texture for those so we can use only 1 pipeline for drawing sprites + shapes
 		const emptyTex = Texture.fromImage(
-			new ImageData(new Uint8ClampedArray([ 255, 255, 255, 255 ]), 1, 1),
+			new ImageData(new Uint8ClampedArray([255, 255, 255, 255]), 1, 1),
+			gl, gc, gopt,
 		)
 
 		if (gopt.background) {
@@ -662,7 +588,9 @@ export default (gopt: KaboomOpt = {}): KaboomCtx => {
 				190, 190, 190, 255,
 				190, 190, 190, 255,
 				128, 128, 128, 255,
-			]), 2, 2), {
+			]), 2, 2),
+			gl, gc, gopt,
+			{
 				wrap: "repeat",
 				filter: "nearest",
 			},
@@ -707,63 +635,6 @@ export default (gopt: KaboomOpt = {}): KaboomCtx => {
 
 	})()
 
-	class SpriteData {
-
-		tex: Texture
-		frames: Quad[] = [ new Quad(0, 0, 1, 1) ]
-		anims: SpriteAnims = {}
-
-		constructor(tex: Texture, frames?: Quad[], anims: SpriteAnims = {}) {
-			this.tex = tex
-			if (frames) this.frames = frames
-			this.anims = anims
-		}
-
-		static from(src: LoadSpriteSrc, opt: LoadSpriteOpt = {}): Promise<SpriteData> {
-			return typeof src === "string"
-				? SpriteData.fromURL(src, opt)
-				: Promise.resolve(SpriteData.fromImage(src, opt),
-				)
-		}
-
-		static fromImage(data: TexImageSource, opt: LoadSpriteOpt = {}): SpriteData {
-			return new SpriteData(
-				Texture.fromImage(data, opt),
-				slice(opt.sliceX || 1, opt.sliceY || 1),
-				opt.anims ?? {},
-			)
-		}
-
-		static fromURL(url: string, opt: LoadSpriteOpt = {}): Promise<SpriteData> {
-			return loadImg(url).then((img) => SpriteData.fromImage(img, opt))
-		}
-
-	}
-
-	class SoundData {
-
-		buf: AudioBuffer
-
-		constructor(buf: AudioBuffer) {
-			this.buf = buf
-		}
-
-		static fromArrayBuffer(buf: ArrayBuffer): Promise<SoundData> {
-			return new Promise((resolve, reject) =>
-				audio.ctx.decodeAudioData(buf, resolve, reject),
-			).then((buf: AudioBuffer) => new SoundData(buf))
-		}
-
-		static fromURL(url: string): Promise<SoundData> {
-			if (isDataURL(url)) {
-				return SoundData.fromArrayBuffer(dataURLToArrayBuffer(url))
-			} else {
-				return fetchArrayBuffer(url).then((buf) => SoundData.fromArrayBuffer(buf))
-			}
-		}
-
-	}
-
 	const audio = (() => {
 
 		// TODO: handle when audio context is unavailable
@@ -790,90 +661,6 @@ export default (gopt: KaboomOpt = {}): KaboomCtx => {
 		}
 
 	})()
-
-	class Asset<D> {
-		done: boolean = false
-		data: D | null = null
-		error: Error | null = null
-		private onLoadEvents: Event<[D]> = new Event()
-		private onErrorEvents: Event<[Error]> = new Event()
-		private onFinishEvents: Event<[]> = new Event()
-		constructor(loader: Promise<D>) {
-			loader.then((data) => {
-				this.data = data
-				this.onLoadEvents.trigger(data)
-			}).catch((err) => {
-				this.error = err
-				if (this.onErrorEvents.numListeners() > 0) {
-					this.onErrorEvents.trigger(err)
-				} else {
-					throw err
-				}
-			}).finally(() => {
-				this.onFinishEvents.trigger()
-				this.done = true
-			})
-		}
-		static loaded<D>(data: D): Asset<D> {
-			const asset = new Asset(Promise.resolve(data))
-			asset.data = data
-			asset.done = true
-			return asset
-		}
-		onLoad(action: (data: D) => void) {
-			this.onLoadEvents.add(action)
-			return this
-		}
-		onError(action: (err: Error) => void) {
-			this.onErrorEvents.add(action)
-			return this
-		}
-		onFinish(action: () => void) {
-			this.onFinishEvents.add(action)
-			return this
-		}
-		then(action: (data: D) => void): Asset<D> {
-			return this.onLoad(action)
-		}
-		catch(action: (err: Error) => void): Asset<D> {
-			return this.onError(action)
-		}
-		finally(action: () => void): Asset<D> {
-			return this.onFinish(action)
-		}
-	}
-
-	class AssetBucket<D> {
-		assets: Map<string, Asset<D>> = new Map()
-		lastUID: number = 0
-		add(name: string | null, loader: Promise<D>): Asset<D> {
-			// if user don't provide a name we use a generated one
-			const id = name ?? (this.lastUID++ + "")
-			const asset = new Asset(loader)
-			this.assets.set(id, asset)
-			return asset
-		}
-		addLoaded(name: string | null, data: D) {
-			const id = name ?? (this.lastUID++ + "")
-			const asset = Asset.loaded(data)
-			this.assets.set(id, asset)
-		}
-		get(handle: string): Asset<D> | void {
-			return this.assets.get(handle)
-		}
-		progress(): number {
-			if (this.assets.size === 0) {
-				return 1
-			}
-			let loaded = 0
-			this.assets.forEach((asset) => {
-				if (asset.done) {
-					loaded++
-				}
-			})
-			return loaded / this.assets.size
-		}
-	}
 
 	const assets = {
 		// prefix for when loading from a url
@@ -998,7 +785,7 @@ export default (gopt: KaboomOpt = {}): KaboomCtx => {
 		return assets.bitmapFonts.add(name, loadImg(src)
 			.then((img) => {
 				return makeFont(
-					Texture.fromImage(img, opt),
+					Texture.fromImage(img, gl, gc, gopt, opt),
 					gw,
 					gh,
 					opt.chars ?? ASCII_CHARS,
@@ -1036,7 +823,7 @@ export default (gopt: KaboomOpt = {}): KaboomCtx => {
 				})
 			}))
 		}
-		return load(SpriteData.from(src).then((atlas) => {
+		return load(SpriteData.from(src, loadImg, gl, gc, gopt, slice).then((atlas) => {
 			const map = {}
 			for (const name in data) {
 				const w = atlas.tex.width
@@ -1074,8 +861,8 @@ export default (gopt: KaboomOpt = {}): KaboomCtx => {
 		return assets.sprites.add(
 			name,
 			typeof src === "string"
-				? SpriteData.fromURL(src, opt)
-				: Promise.resolve(SpriteData.fromImage(src, opt),
+				? SpriteData.fromURL(src, loadImg, gl, gc, gopt, slice, opt)
+				: Promise.resolve(SpriteData.fromImage(src, gl, gc, gopt, slice, opt),
 				),
 		)
 	}
@@ -1183,8 +970,8 @@ export default (gopt: KaboomOpt = {}): KaboomCtx => {
 		return assets.sounds.add(
 			name,
 			typeof src === "string"
-				? SoundData.fromURL(src)
-				: SoundData.fromArrayBuffer(src),
+				? SoundData.fromURL(src, audio, fetchArrayBuffer)
+				: SoundData.fromArrayBuffer(src, audio),
 		)
 	}
 
@@ -1288,8 +1075,7 @@ export default (gopt: KaboomOpt = {}): KaboomCtx => {
 		| BitmapFontData
 		| Asset<BitmapFontData>
 		| string
-		| void
-	{
+		| void {
 		if (!src) {
 			return resolveFont(gopt.font ?? DEF_FONT)
 		}
@@ -2272,7 +2058,7 @@ export default (gopt: KaboomOpt = {}): KaboomCtx => {
 
 		if (opt.outline) {
 			drawLines({
-				pts: [ ...opt.pts, opt.pts[0] ],
+				pts: [...opt.pts, opt.pts[0]],
 				radius: opt.radius,
 				width: opt.outline.width,
 				color: opt.outline.color,
@@ -2441,9 +2227,11 @@ export default (gopt: KaboomOpt = {}): KaboomCtx => {
 
 			const atlas: FontAtlas = fontAtlases[fontName] ?? {
 				font: {
-					tex: new Texture(FONT_ATLAS_SIZE, FONT_ATLAS_SIZE, {
-						filter: "linear",
-					}),
+					tex: new Texture(FONT_ATLAS_SIZE, FONT_ATLAS_SIZE,
+						gl, gc, gopt,
+						{
+							filter: "linear",
+						}),
 					map: {},
 					size: DEF_TEXT_CACHE_SIZE,
 				},
@@ -3066,8 +2854,8 @@ export default (gopt: KaboomOpt = {}): KaboomCtx => {
 			return app.numKeyDown > 0
 		} else {
 			return app.keyStates[k] === "pressed"
-			|| app.keyStates[k] === "rpressed"
-			|| app.keyStates[k] === "down"
+				|| app.keyStates[k] === "rpressed"
+				|| app.keyStates[k] === "down"
 		}
 	}
 
@@ -4011,7 +3799,7 @@ export default (gopt: KaboomOpt = {}): KaboomCtx => {
 	function follow(obj: GameObj, offset?: Vec2): FollowComp {
 		return {
 			id: "follow",
-			require: [ "pos" ],
+			require: ["pos"],
 			follow: {
 				obj: obj,
 				offset: offset ?? vec2(0),
@@ -4033,7 +3821,7 @@ export default (gopt: KaboomOpt = {}): KaboomCtx => {
 		const d = typeof dir === "number" ? Vec2.fromAngle(dir) : dir.unit()
 		return {
 			id: "move",
-			require: [ "pos" ],
+			require: ["pos"],
 			update(this: GameObj<PosComp>) {
 				this.move(d.scale(speed))
 			},
@@ -4047,7 +3835,7 @@ export default (gopt: KaboomOpt = {}): KaboomCtx => {
 		let isOut = false
 		return {
 			id: "offscreen",
-			require: [ "pos" ],
+			require: ["pos"],
 			isOffScreen(this: GameObj<PosComp>): boolean {
 				const pos = toScreen(this.pos)
 				const screenRect = new Rect(vec2(0), width(), height())
@@ -4186,7 +3974,7 @@ export default (gopt: KaboomOpt = {}): KaboomCtx => {
 					return null
 				}
 				// if (this.colliding[other.id]) {
-					// return this.colliding[other.id]
+				// return this.colliding[other.id]
 				// }
 				const a1 = this.worldArea()
 				const a2 = other.worldArea()
@@ -4518,7 +4306,7 @@ export default (gopt: KaboomOpt = {}): KaboomCtx => {
 						loop: false,
 						pingpong: false,
 						speed: 0,
-						onEnd: () => {},
+						onEnd: () => { },
 					}
 					: {
 						name: name,
@@ -4526,7 +4314,7 @@ export default (gopt: KaboomOpt = {}): KaboomCtx => {
 						loop: opt.loop ?? anim.loop ?? false,
 						pingpong: opt.pingpong ?? anim.pingpong ?? false,
 						speed: opt.speed ?? anim.speed ?? 10,
-						onEnd: opt.onEnd ?? (() => {}),
+						onEnd: opt.onEnd ?? (() => { }),
 					}
 
 				this.frame = typeof anim === "number"
@@ -4779,7 +4567,7 @@ export default (gopt: KaboomOpt = {}): KaboomCtx => {
 		return {
 
 			id: "body",
-			require: [ "pos", "area" ],
+			require: ["pos", "area"],
 			jumpForce: opt.jumpForce ?? DEF_JUMP_FORCE,
 			gravityScale: opt.gravityScale ?? 1,
 			isStatic: opt.isStatic ?? false,
@@ -4951,7 +4739,7 @@ export default (gopt: KaboomOpt = {}): KaboomCtx => {
 		const events = []
 		return {
 			id: "doubleJump",
-			require: [ "body" ],
+			require: ["body"],
 			numJumps: numJumps,
 			add(this: GameObj<BodyComp | DoubleJumpComp>) {
 				events.push(this.onGround(() => {
@@ -5155,7 +4943,7 @@ export default (gopt: KaboomOpt = {}): KaboomCtx => {
 			},
 
 			update() {
-			// execute the enter event for initState
+				// execute the enter event for initState
 				if (!didFirstEnter) {
 					trigger("enter", initState)
 					didFirstEnter = true
@@ -5179,7 +4967,7 @@ export default (gopt: KaboomOpt = {}): KaboomCtx => {
 		let t = 0
 		let done = false
 		return {
-			require: [ "opacity" ],
+			require: ["opacity"],
 			add(this: GameObj<OpacityComp>) {
 				this.opacity = 0
 			},
@@ -5512,7 +5300,7 @@ export default (gopt: KaboomOpt = {}): KaboomCtx => {
 		let time = 0
 		return {
 			id: "boom",
-			require: [ "scale" ],
+			require: ["scale"],
 			update(this: GameObj<ScaleComp>) {
 				const s = Math.sin(time * speed) * size
 				if (s < 0) {
@@ -5568,42 +5356,6 @@ export default (gopt: KaboomOpt = {}): KaboomCtx => {
 
 	const DEF_HASH_GRID_SIZE = 64
 
-	class Collision {
-		source: GameObj
-		target: GameObj
-		displacement: Vec2
-		resolved: boolean = false
-		constructor(source: GameObj, target: GameObj, dis: Vec2, resolved = false) {
-			this.source = source
-			this.target = target
-			this.displacement = dis
-			this.resolved = resolved
-		}
-		reverse() {
-			return new Collision(
-				this.target,
-				this.source,
-				this.displacement.scale(-1),
-				this.resolved,
-			)
-		}
-		isLeft() {
-			return this.displacement.x > 0
-		}
-		isRight() {
-			return this.displacement.x < 0
-		}
-		isTop() {
-			return this.displacement.y > 0
-		}
-		isBottom() {
-			return this.displacement.y < 0
-		}
-		preventResolve() {
-			this.resolved = true
-		}
-	}
-
 	function checkFrame() {
 
 		// TODO: persistent grid?
@@ -5646,10 +5398,10 @@ export default (gopt: KaboomOpt = {}): KaboomCtx => {
 				// insert & check against all covered grids
 				for (let x = xmin; x <= xmax; x++) {
 					for (let y = ymin; y <= ymax; y++) {
-						if(!grid[x]) {
+						if (!grid[x]) {
 							grid[x] = {}
 							grid[x][y] = [aobj]
-						} else if(!grid[x][y]) {
+						} else if (!grid[x][y]) {
 							grid[x][y] = [aobj]
 						} else {
 							const cell = grid[x][y]
